requirements.txt
pandas
numpy
scikit-learn
xgboost
prophet
matplotlib
seaborn
plotly
joblib
pyyaml
streamlit
pytest